Перенос консольных команд Door в другой класс.
